requirements.txt
celery==5.0.5
channels==3.0.3
channels-redis==3.2.0
Django==3.1.6
django-constance==2.8.0
django-cors-headers==3.2.1
django-filter==2.4.0
django-storages==1.9.1
djangorestframework==3.12.4
drf-yasg==1.20.0
fabric==2.5.0
gevent==21.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-decouple==3.3
python-dateutil==2.8.2
redis==3.5.3
supervisor==4.2.1
whitenoise==5.1.0
